Core: Export every XDP action from pythonbpf.helper
helpers.py defined all five and return_utils mapped them, but only
XDP_DROP and XDP_PASS were exported, so the xdp_tx port had to import
XDP_TX from vmlinux and live under vmlinux/, skipped on any host without
a generated module. It imports from pythonbpf.helper and lives under
xdp/ now, returning XDP_TX bare like its siblings and the C original,
which is the shape the return fast path resolves without vmlinux.

+++ tests/kernel_selftest_equivalent/xdp/xdp_tx.py
@@ -3,18 +3,15 @@
 # Bounce every packet back out of the interface it arrived on. Upstream is
 # the transmit side of the veth XDP tests.
 #
-# XDP_TX comes from vmlinux because pythonbpf.helper exports only XDP_PASS
-# and XDP_DROP; that is why this lives under vmlinux/.
-
 from pythonbpf import bpf, section, bpfglobal, compile
-from vmlinux import XDP_TX
+from pythonbpf.helper import XDP_TX
 from ctypes import c_void_p, c_int64
 
 
 @bpf
 @section("xdp")
 def xdp_tx(xdp: c_void_p) -> c_int64:
-    return c_int64(XDP_TX)
+    return XDP_TX  # bare, like the C: resolved by the return fast path
 
 
 @bpf
